pyproject.toml
[project]
name = "arcgentica"
version = "0.1.0"
description = "An AI agent system for solving ARC-AGI challenges using LLMs based on Agentica, inspired by Poetiq"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "datasets>=4.4.1",
    "symbolica-agentica==0.4.0",
    "asynciolimiter==1.2.0",
    "litellm==1.78.2",
    "numpy",
    "scipy",
    "sympy",
    "python-dotenv",
    "scikit-image",
    "matplotlib>=3.10.8",
    "adjusttext>=1.3.0",
    "openai>=2.16.0",
]
